Added batch files. Renamed Base class to FrameworkUtil. Assigned flash server port in default.properties. Updated ReadMe.md. Tests executed.
