+++ batch_files/jar_maker.bat
@@ -1,7 +1,6 @@
 @echo off
 echo Making MaxSoft IntelliAPI JAR .............
 
-cd ..
 cd .. 
 
 mvn clean install -DskipTests
