+++ specs/other/Text Files Operations.spec
@@ -31,11 +31,11 @@ tags: text_file, regression
    |#gender                         |yes                |Scenario       |varGender               |                        |
    |#status                         |yes                |Scenario       |varStatus               |                        |
 * And the user set the request authentication configurations as follows
-   |Configuration                                                     |Configuration Value                                                   |
-   |------------------------------------------------------------------|----------------------------------------------------------------------|
-   |Is authentication required?                                       |Yes                                                                   |
-   |Do you need to retrieve the access token from the text file?      |No                                                                    |
-   |Provide the access token if you need to authorize the API manually|<file:./src/test/resources/access_tokens/saved_valid_access_token.txt>|
+   |Configuration                                                     |Configuration Value                                  |
+   |------------------------------------------------------------------|-----------------------------------------------------|
+   |Is authentication required?                                       |Yes                                                  |
+   |Do you need to retrieve the access token from the text file?      |No                                                   |
+   |Provide the access token if you need to authorize the API manually|<file:./src/test/resources/payloads/access_token.txt>|
 * When the user invokes the API
 * Then the status code for the request is "200"
 * And save the JSON Path values in the response inside the data stores
