+++ src/main/java/com/maxsoft/intelliapi/stepimpl/database/MongoStepImpl.java
@@ -8,7 +8,7 @@ import com.thoughtworks.gauge.TableRow;
 
 import java.util.List;
 
-import static com.maxsoft.intelliapi.common.Base.*;
+import static com.maxsoft.intelliapi.util.FrameworkUtil.*;
 import static com.maxsoft.intelliapi.common.Constants.*;
 import static com.maxsoft.intelliapi.util.DataStoreProcessor.*;
