+++ src/test/resources/payloads/access_token.txt
@@ -1,0 +1,1 @@
+Bearer 5aa8220420419fd5890bb88a1767ed7cb1abc4412024bfff91513a40d6e19823
